points: Fix styling.
